ebdjango/requirements.txt
asgiref==3.5.1
boto3==1.22.13
botocore==1.25.13
certifi==2021.10.8
cffi==1.15.0
cryptography==37.0.2
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
jmespath==1.0.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
wincertstore==0.2
